Migrate balance info to pull from table

+++ server/BudgetBoard-Database/Models/Goal.cs
@@ -5,9 +5,9 @@ public class Goal
     public Guid ID { get; set; }
     public string Name { get; set; } = string.Empty;
     public DateTime? CompleteDate { get; set; }
-    public float Amount { get; set; } = 0.0f;
-    public float? InitialAmount { get; set; }
-    public float? MonthlyContribution { get; set; }
+    public decimal Amount { get; set; } = 0.0M;
+    public decimal? InitialAmount { get; set; }
+    public decimal? MonthlyContribution { get; set; }
     public ICollection<Account> Accounts { get; set; } = new List<Account>();
     public required Guid UserID { get; set; }
     public ApplicationUser? User { get; set; } = null!;
@@ -17,8 +17,8 @@ public class NewGoal
 {
     public required string Name { get; set; }
     public DateTime? CompleteDate { get; set; }
-    public required float Amount { get; set; }
-    public float? InitialAmount { get; set; }
-    public float? MonthlyContribution { get; set; }
+    public required decimal Amount { get; set; }
+    public decimal? InitialAmount { get; set; }
+    public decimal? MonthlyContribution { get; set; }
     public required ICollection<string> AccountIds { get; set; }
 }
